Code modified
Code modified

+++ spring/SpringJPARepoDistributedTransaction/src/main/java/com/ashish/credit/repositories/CreditCardTransactionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ashish.credit.entities.Transaction;
 
-public interface CreditCardTransactionRepository extends JpaRepository<Transaction, Long> {
+public interface CreditCardTransactionRepository extends JpaRepository<Transaction, Integer> {
 
 }
